+++ src/main/java/com/hyunn/commerceplatform/entity/DiscountCodes.java
@@ -1,5 +1,6 @@
 package com.hyunn.commerceplatform.entity;
 
+import com.hyunn.commerceplatform.entity.types.DiscountType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
