+++ src/main/java/com/hyunn/commerceplatform/entity/OrderItems.java
@@ -1,5 +1,6 @@
 package com.hyunn.commerceplatform.entity;
 
+import com.hyunn.commerceplatform.entity.types.OrderStatus;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
