+++ src/main/java/com/hyunn/commerceplatform/entity/Terms.java
@@ -1,5 +1,6 @@
 package com.hyunn.commerceplatform.entity;
 
+import com.hyunn.commerceplatform.entity.types.TermType;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
